Reorganize dataset inheritance
CpVtonDataset is now an abstract base class that represents how data should be loaded

Viton, VVT, and MPV subclass CpVton to change the specific paths
