+++ datasets/cpvton_dataset.py
@@ -1,4 +1,5 @@
 # coding=utf-8
+from abc import ABC, abstractmethod
 import torch
 import torch.utils.data as data
 import torchvision.transforms as transforms
@@ -13,11 +14,12 @@ import json
 from datasets.util import segment_cloths_from_image
 
 
-class CPDataset(data.Dataset):
-    """Dataset for CP-VTON. """
+class CpVtonDataset(ABC, data.Dataset):
+    """ Loads all the necessary items for CP-Vton """
 
     def __init__(self, opt):
-        super(CPDataset, self).__init__()
+        super(CpVtonDataset, self).__init__()
+        self.CLOTH_THRESH = 240
         # base setting
         self.opt = opt
         self.root = opt.dataroot
@@ -36,25 +38,17 @@ class CPDataset(data.Dataset):
                 transforms.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5)),
             ]
         )
-
+        self.image_names = []
         # load data list
         self.load_file_paths()
 
+    @abstractmethod
     def load_file_paths(self):
         """
         Reads the datalist txt file for CP-VTON
         sets self.image_names and self.cloth_names. they should correspond 1-to-1
         """
-        im_names = []
-        c_names = []
-        with open(osp.join(self.opt.dataroot, self.opt.data_list), "r") as f:
-            for line in f.readlines():
-                im_name, c_name = line.strip().split()
-                im_names.append(im_name)
-                c_names.append(c_name)
-
-        self.image_names = im_names
-        self.cloth_names = c_names
+        pass
 
     def __len__(self):
         return len(self.image_names)
@@ -70,8 +64,17 @@ class CPDataset(data.Dataset):
         :return: cloth, cloth_mask
         """
         cloth = self.get_input_cloth(index)
-        cloth_mask = self.get_input_cloth_mask(index)
+        cloth_mask = self.get_input_cloth_mask(cloth)
         return cloth, cloth_mask
+
+    def get_input_cloth_mask(self, input_cloth: torch.Tensor):
+        """ Creates a mask directly from the input_cloth """
+        # make the mask
+        low = torch.zeros_like(input_cloth)
+        high = torch.ones_like(input_cloth)
+        cloth_mask = torch.where(input_cloth >= self.CLOTH_THRESH, low, high)
+        cloth_mask = cloth_mask[0].unsqueeze(0)  # the mask should be a single channel
+        return cloth_mask
 
     def get_input_cloth(self, index):
         """
@@ -83,35 +86,18 @@ class CPDataset(data.Dataset):
         c = self.to_tensor_and_norm(c)  # [-1,1]
         return c
 
+    @abstractmethod
     def get_input_cloth_path(self, index):
         """
         Get the file path for the product image input.
         Called by get_input_cloth()
         """
-        c_name = self.get_input_cloth_name(index)
-        folder = "cloth" if self.stage == "GMM" else "warp-cloth"
-        cloth_path = osp.join(self.data_path, folder, c_name)
-        return cloth_path
+        pass
 
+    @abstractmethod
     def get_input_cloth_name(self, index):
         # determines the written thing
-        return self.cloth_names[index]
-
-    def get_input_cloth_mask(self, index):
-        """from cp-vton"""
-        mask_path = self.get_input_cloth_mask_path(index)
-        cm = Image.open(mask_path)
-        cm_array = np.array(cm)
-        cm_array = (cm_array >= 128).astype(np.float32)
-        cm = torch.from_numpy(cm_array)  # [0,1]
-        cm.unsqueeze_(0)
-        return cm
-
-    def get_input_cloth_mask_path(self, index):
-        c_name = self.get_input_cloth_name(index)
-        folder = "cloth-mask" if self.stage == "GMM" else "warp-mask"
-        mask_path = osp.join(self.data_path, folder, c_name)
-        return mask_path
+        pass
 
     ########################
     # PERSON REPRESENTATION
@@ -155,15 +141,6 @@ class CPDataset(data.Dataset):
         im = self.to_tensor_and_norm(im)  # [-1,1]
         return im
 
-    def get_person_image_name(self, index):
-        """ basename of the image file """
-        return self.image_names[index]
-
-    def get_person_image_path(self, index):
-        im_name = self.get_person_image_name(index)
-        image_path = osp.join(self.data_path, "image", im_name)
-        return image_path
-
     def get_person_parsed(self, index):
         """ loads parsing image """
         parsed_path = self.get_person_parsed_path(index)
@@ -172,12 +149,34 @@ class CPDataset(data.Dataset):
         parse_array = np.array(im_parse)
         return parse_array
 
-    def get_person_parsed_path(self, index):
-        """ path of the clothing seguemtnation """
-        im_name = self.get_person_image_name(index)
-        parse_name = im_name.replace(".jpg", ".png")
-        parsed_path = osp.join(self.data_path, "image-parse", parse_name)
-        return parsed_path
+    def get_input_person_head(self, im, _parse_array):
+        """ from cp-vton, get the floating head alone"""
+        # ISOLATE HEAD. head parts, probably face, hair, sunglasses. combines into a 1d binary mask
+        _parse_head = (
+            (_parse_array == 1).astype(np.float32)
+            + (_parse_array == 2).astype(np.float32)
+            + (_parse_array == 4).astype(np.float32)
+            + (_parse_array == 13).astype(np.float32)
+        )
+        _phead = torch.from_numpy(_parse_head)  # [0,1]
+        im_h = im * _phead - (1 - _phead)  # [-1,1], fill 0 for other parts
+        return im_h
+
+    def get_input_person_body_silhouette(self, _parse_array):
+        """ from cp-vton, the body silhouette """
+        # ISOLATE BODY SHAPE
+        # removes the background
+        _parse_shape = (_parse_array > 0).astype(np.float32)
+        # silhouette downsample, reduces resolution, makes the silhouette "blurry"
+        _parse_shape = Image.fromarray((_parse_shape * 255).astype(np.uint8))
+        _parse_shape = _parse_shape.resize(
+            (self.fine_width // 16, self.fine_height // 16), Image.BILINEAR
+        )
+        _parse_shape = _parse_shape.resize(
+            (self.fine_width, self.fine_height), Image.BILINEAR
+        )
+        silhouette = self.to_tensor_and_norm(_parse_shape)  # [-1,1]
+        return silhouette
 
     def get_input_person_pose(self, index):
         """from cp-vton, loads the pose as white squares
@@ -197,13 +196,6 @@ class CPDataset(data.Dataset):
         pose_map, im_pose = self.convert_pose_data_to_pose_map_and_vis(pose_data)
 
         return pose_map, im_pose
-
-    def get_input_person_pose_path(self, index):
-        """ path to pose keypoints """
-        im_name = self.get_person_image_name(index)
-        _pose_name = im_name.replace(".jpg", "_keypoints.json")
-        pose_path = osp.join(self.data_path, "pose", _pose_name)
-        return pose_path
 
     def convert_pose_data_to_pose_map_and_vis(self, pose_data):
         """
@@ -251,34 +243,28 @@ class CPDataset(data.Dataset):
         im_pose = self.to_tensor_and_norm(im_pose)
         return pose_map, im_pose
 
-    def get_input_person_head(self, im, _parse_array):
-        """ from cp-vton, get the floating head alone"""
-        # ISOLATE HEAD. head parts, probably face, hair, sunglasses. combines into a 1d binary mask
-        _parse_head = (
-            (_parse_array == 1).astype(np.float32)
-            + (_parse_array == 2).astype(np.float32)
-            + (_parse_array == 4).astype(np.float32)
-            + (_parse_array == 13).astype(np.float32)
-        )
-        _phead = torch.from_numpy(_parse_head)  # [0,1]
-        im_h = im * _phead - (1 - _phead)  # [-1,1], fill 0 for other parts
-        return im_h
+    @abstractmethod
+    def get_person_image_path(self, index):
+        pass
 
-    def get_input_person_body_silhouette(self, _parse_array):
-        """ from cp-vton, the body silhouette """
-        # ISOLATE BODY SHAPE
-        # removes the background
-        _parse_shape = (_parse_array > 0).astype(np.float32)
-        # silhouette downsample, reduces resolution, makes the silhouette "blurry"
-        _parse_shape = Image.fromarray((_parse_shape * 255).astype(np.uint8))
-        _parse_shape = _parse_shape.resize(
-            (self.fine_width // 16, self.fine_height // 16), Image.BILINEAR
-        )
-        _parse_shape = _parse_shape.resize(
-            (self.fine_width, self.fine_height), Image.BILINEAR
-        )
-        silhouette = self.to_tensor_and_norm(_parse_shape)  # [-1,1]
-        return silhouette
+    @abstractmethod
+    def get_person_image_name(self, index):
+        """ basename of the image file """
+        pass
+
+    @abstractmethod
+    def get_person_parsed_path(self, index):
+        """ path of the clothing seguemtnation """
+        pass
+
+    @abstractmethod
+    def get_input_person_pose_path(self, index):
+        """ path to pose keypoints """
+        pass
+
+    ########################
+    # getitem
+    ########################
 
     def __getitem__(self, index):
         im_name = self.get_person_image_name(index)
@@ -304,32 +290,34 @@ class CPDataset(data.Dataset):
         else:
             im_grid = ""
 
-        assert cloth.shape == torch.Size(
-            [3, 256, 192]
-        ), f"cloth.shape = {cloth.shape} on {im_path} {cloth_path}"
-        assert cloth_mask.shape == torch.Size(
-            [1, 256, 192]
-        ), f"cloth_mask.shape = {cloth_mask.shape} on {im_path} {cloth_path}"
-        assert im.shape == torch.Size([3, 256, 192]), f"im = {im.shape} on {im_path}"
-        assert im_cloth.shape == torch.Size(
-            [3, 256, 192]
-        ), f"im_cloth = {im_cloth} on {im_name}"
-        assert im_pose.shape == torch.Size(
-            [1, 256, 192]
-        ), f"im_pose.shape = {im_pose.shape} on {im_path}"
-        assert silhouette.shape == torch.Size(
-            [1, 256, 192]
-        ), f"silhouette.shape = {silhouette.shape} on {im_path}"
-        assert im_head.shape == torch.Size(
-            [3, 256, 192]
-        ), f"im_head = {im_head} on {im_path}"
-        assert agnostic.shape == torch.Size(
-            [22, 256, 192]
-        ), f"agnostic = {agnostic} on {im_path}"
-        if im_grid != "":
-            assert im_grid.shape == torch.Size(
+        def run_assertions():
+            assert cloth.shape == torch.Size(
                 [3, 256, 192]
-            ), f"im_grid = {im_grid} on {im_path}"
+            ), f"cloth.shape = {cloth.shape} on {im_path} {cloth_path}"
+            assert cloth_mask.shape == torch.Size(
+                [1, 256, 192]
+            ), f"cloth_mask.shape = {cloth_mask.shape} on {im_path} {cloth_path}"
+            assert im.shape == torch.Size([3, 256, 192]), f"im = {im.shape} on {im_path}"
+            assert im_cloth.shape == torch.Size(
+                [3, 256, 192]
+            ), f"im_cloth = {im_cloth} on {im_name}"
+            assert im_pose.shape == torch.Size(
+                [1, 256, 192]
+            ), f"im_pose.shape = {im_pose.shape} on {im_path}"
+            assert silhouette.shape == torch.Size(
+                [1, 256, 192]
+            ), f"silhouette.shape = {silhouette.shape} on {im_path}"
+            assert im_head.shape == torch.Size(
+                [3, 256, 192]
+            ), f"im_head = {im_head} on {im_path}"
+            assert agnostic.shape == torch.Size(
+                [22, 256, 192]
+            ), f"agnostic = {agnostic} on {im_path}"
+            if im_grid is not "":
+                assert im_grid.shape == torch.Size(
+                    [3, 256, 192]
+                ), f"im_grid = {im_grid} on {im_path}"
+        run_assertions()
 
         result = {
             "dataset_name": self.__class__.__name__,
@@ -380,34 +368,3 @@ class CPDataLoader(object):
         return batch
 
 
-if __name__ == "__main__":
-    print("Check the dataset for geometric matching module!")
-
-    import argparse
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--dataroot", default="data")
-    parser.add_argument("--datamode", default="train")
-    parser.add_argument("--stage", default="GMM")
-    parser.add_argument("--data_list", default="train_pairs.txt")
-    parser.add_argument("--fine_width", type=int, default=192)
-    parser.add_argument("--fine_height", type=int, default=256)
-    parser.add_argument("--radius", type=int, default=3)
-    parser.add_argument("--shuffle", action="store_true", help="shuffle input data")
-    parser.add_argument("-b", "--batch-size", type=int, default=4)
-    parser.add_argument("-j", "--workers", type=int, default=1)
-
-    opt = parser.parse_args()
-    dataset = CPDataset(opt)
-    data_loader = CPDataLoader(opt, dataset)
-
-    print(
-        "Size of the dataset: %05d, dataloader: %04d"
-        % (len(dataset), len(data_loader.data_loader))
-    )
-    first_item = dataset.__getitem__(0)
-    first_batch = data_loader.next_batch()
-
-    from IPython import embed
-
-    embed()
